requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.2
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.1.0
colorama==0.4.6
discord.py==2.4.0
distro==1.9.0
dnspython==2.3.0
exceptiongroup==1.2.0
ffmpeg==1.4
ffmpeg-python==0.2.0
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
multidict==6.0.4
mutagen==1.47.0
openai==0.28.0
outcome==1.3.0.post0
pafy==0.5.5
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.3.3
PyNaCl==1.5.0
PySocks==1.7.1
requests==2.32.3
selenium==4.22.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
yarl==1.8.2
youtube-dl==2021.12.17
yt-dlp==2024.7.9
yt-dlp-youtube-oauth2 @ https://github.com/coletdjnz/yt-dlp-youtube-oauth2/archive/refs/heads/master.zip
